requirements.txt
asgiref==3.6.0
black==23.3.0
boltons==23.0.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-bootstrap5==23.3
django-charid-field==0.3
django-csvimport==2.17
django-extensions==3.2.1
django-filter==23.2
django_heroku
django-import-export==3.2.0
django-tables2==2.5.3
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
flake8==6.0.0
flake8-black==0.3.6
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
json-table-schema==0.2.1
jsonpatch==1.32
jsonpointer==2.0
lxml==4.9.2
MarkupPy==1.14
mccabe==0.7.0
messytables==0.15.2
mypy-extensions==1.0.0
numpy==1.24.3
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
pip==23.1.2
platformdirs==3.5.1
pluggy==1.0.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycosat==0.6.3
pycparser==2.21
pyflakes==3.0.1
pyOpenSSL==23.2.0
PySocks==1.7.1
pytest==7.3.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-magic==0.4.27
python-ulid==1.1.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
setuptools==67.7.2
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
tablib==3.4.0
tabulate==0.9.0
tomli==2.0.1
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
ulid==1.1
urllib3==2.0.2
webencodings==0.5.1
wheel==0.40.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.15.0
zstandard==0.19.0
